% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_antsExtractXptrAsString.R
\docType{methods}
\name{.int_antsExtractXptrAsString,numeric-method}
\alias{.int_antsExtractXptrAsString,numeric-method}
\title{#' @aliases .int_antsExtractXptrAsString,antsMatrix-method
setMethod(f = ".int_antsExtractXptrAsString", signature = c("antsMatrix"), definition = function(object) {
  return(as.character(c(object@pointer)))
})}
\usage{
\S4method{.int_antsExtractXptrAsString}{numeric}(object)
}
\description{
#' @aliases .int_antsExtractXptrAsString,antsMatrix-method
setMethod(f = ".int_antsExtractXptrAsString", signature = c("antsMatrix"), definition = function(object) {
  return(as.character(c(object@pointer)))
})
}
